another travis try

+++ Travis-ci/install-zef.sh
@@ -1,13 +1,25 @@
 #!/usr/bin/env sh
 
-set -ev
+set -v
 
 # Install zef
 
-git clone https://github.com/ugexe/zef.git
-cd zef
-perl6 -Ilib bin/zef install .
-cd ..
+/bin/ls -lR
+
+cd ${TRAVIS_BUILD_DIR}/Travis-ci
+
+if [ ! -e zef ]
+then
+
+  git clone https://github.com/ugexe/zef.git
+  cd zef
+  perl6 -Ilib bin/zef install .
+
+else
+
+  zef update
+  
+fi
 
 exit 0
